backend/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
firebase-admin==6.7.0
django-cors-headers==4.6.0
PyJWT==2.10.1
bcrypt==4.2.1
gunicorn==23.0.0
python-dotenv==1.0.1
